use parser ik (#30)
<!--
Thank you for contributing to OceanBase! 
Please feel free to ping the maintainers for the review!
-->

## Summary
<!-- 
Please clearly and concisely describe the purpose of this pull request.
If this pull request resolves an issue, please link it via "close #xxx"
or "fix #xxx".
-->

parser ik is a better tokenizer than space when using chinese

## Solution Description
<!-- Please clearly and concisely describe your solution. -->